Unified generators to proxy through a single class.
