+++ src/Helpers/GeneratedContentHelper.php
@@ -2,15 +2,9 @@
 
 namespace Drupal\generated_content\Helpers;
 
-use Drupal\Component\Render\FormattableMarkup;
-use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
 use Drupal\Core\Url;
-use Drupal\field\Entity\FieldConfig;
-use Drupal\generated_content\GeneratedContentRepository;
-use Drupal\generated_content\Generator\GeneratedContentRandom;
 use Drupal\menu_link_content\Entity\MenuLinkContent;
 use Drupal\taxonomy\Entity\Term;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Class GeneratedContentHelper.
@@ -19,93 +13,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *
  * @package \Drupal\generated_content
  */
-class GeneratedContentHelper implements ContainerInjectionInterface {
-
-  use GeneratedContentVariationTrait;
-
-  /**
-   * The helper singleton.
-   *
-   * @var \Drupal\generated_content\Helpers\GeneratedContentHelper
-   */
-  protected static $instance = NULL;
-
-  /**
-   * The repository singleton.
-   *
-   * @var \Drupal\generated_content\GeneratedContentRepository
-   */
-  protected static $repository = NULL;
-
-  /**
-   * Use verbose mode.
-   *
-   * @var bool
-   */
-  protected static $verbose = TRUE;
-
-  /**
-   * GeneratedContentHelper constructor.
-   */
-  public function __construct(GeneratedContentRepository $repository) {
-    static::$repository = $repository;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      GeneratedContentRepository::getInstance()
-    );
-  }
-
-  /**
-   * Get this helper instance.
-   *
-   * @return \Drupal\generated_content\Helpers\GeneratedContentHelper
-   *   The repository.
-   */
-  public static function getInstance() {
-    if (!self::$instance) {
-      static::$instance = \Drupal::service('class_resolver')
-        ->getInstanceFromDefinition(static::class);
-    }
-
-    return self::$instance;
-  }
-
-  /**
-   * Log verbose progress.
-   */
-  public static function log() {
-    if (self::$verbose) {
-      if (function_exists('drush_print')) {
-        // Strip all tags, but still decode some HTML entities.
-        drush_print(html_entity_decode(strip_tags(call_user_func_array('sprintf', func_get_args()))));
-      }
-      else {
-        // Support HTML, but still use plain strings for simplicity.
-        \Drupal::messenger()->addMessage(new FormattableMarkup(call_user_func_array('sprintf', func_get_args()), []));
-      }
-    }
-  }
-
-  /**
-   * Add entities to the repository.
-   *
-   * Useful to make entities available within the creation callback before
-   * they are returned to allow referencing within the same callback.
-   * For example, to create nodes and fill-in related nodes in the follow-up
-   * iterations.
-   *
-   * @param mixed $entities
-   *   Array of entities.
-   */
-  public static function addToRepository($entities) {
-    $entities = is_array($entities) ? $entities : [$entities];
-    self::$repository->addEntitiesNoTracking($entities);
-  }
+class GeneratedContentHelper extends GeneratedContentAbstractHelper {
 
   /**
    * Select a random user.
@@ -117,7 +25,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
     $users = [1 => 1];
     $users += self::$repository->getEntities('user', 'user');
 
-    return GeneratedContentRandom::arrayItem($users);
+    return self::randomArrayItem($users);
   }
 
   /**
@@ -138,7 +46,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
       $nodes = array_shift($nodes);
     }
 
-    return GeneratedContentRandom::arrayItem($nodes);
+    return self::randomArrayItem($nodes);
   }
 
   /**
@@ -168,7 +76,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
       $nodes = $filtered_nodes;
     }
 
-    return $count ? GeneratedContentRandom::arrayItems($nodes, $count) : $nodes;
+    return $count ? self::randomArrayItems($nodes, $count) : $nodes;
   }
 
   /**
@@ -180,15 +88,15 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
    *   Optional term count to return. If NULL - all terms will be returned.
    *   If specified - this count of already randomised terms will be returned.
    *
-   * @return \Drupal\taxonomy\Entity\Term[]
+   * @return \Drupal\Core\Entity\EntityInterface[]
    *   Array of terms.
    */
   public static function randomRealTerms($vid, $count = NULL) {
-    $terms = \Drupal::service('entity_type.manager')
+    $terms = self::$entityTypeManager
       ->getStorage('taxonomy_term')
       ->loadByProperties(['vid' => $vid]);
 
-    return $count ? GeneratedContentRandom::arrayItems($terms, $count) : $terms;
+    return $count ? self::randomArrayItems($terms, $count) : $terms;
   }
 
   /**
@@ -197,7 +105,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
    * @param string $vid
    *   Vocabulary machine name.
    *
-   * @return \Drupal\taxonomy\Entity\Term
+   * @return \Drupal\Core\Entity\EntityInterface[]|false
    *   The term.
    */
   public static function randomRealTerm($vid) {
@@ -215,13 +123,13 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
    *   Optional term count to return. If NULL - all terms will be returned.
    *   If specified - this count of already randomised terms will be returned.
    *
-   * @return \Drupal\taxonomy\Entity\Term[]
+   * @return \Drupal\Core\Entity\EntityInterface[]
    *   Array of terms.
    */
   public static function randomTerms($vid, $count = NULL) {
     $terms = self::$repository->getEntities('taxonomy_term', $vid);
 
-    return $count ? GeneratedContentRandom::arrayItems($terms, $count) : $terms;
+    return $count ? self::randomArrayItems($terms, $count) : $terms;
   }
 
   /**
@@ -230,7 +138,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
    * @param string $vid
    *   Vocabulary machine name.
    *
-   * @return \Drupal\taxonomy\Entity\Term
+   * @return \Drupal\Core\Entity\EntityInterface[]|false
    *   The term.
    */
   public static function randomTerm($vid) {
@@ -258,14 +166,14 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
   public static function randomFieldAllowedValues($entity_type, $bundle, $field_name, $count = NULL) {
     $allowed_values = [];
 
-    $field_info = FieldConfig::loadByName($entity_type, $bundle, $field_name);
+    $field_info = self::$entityTypeManager->getStorage('field_config')->load($entity_type . '.' . $bundle . '.' . $field_name);
     if ($field_info) {
       $allowed_values = $field_info->getFieldStorageDefinition()->getSetting('allowed_values');
     }
 
     $allowed_values = array_keys($allowed_values);
 
-    return $count ? GeneratedContentRandom::arrayItems($allowed_values, $count) : $allowed_values;
+    return $count ? self::randomArrayItems($allowed_values, $count) : $allowed_values;
   }
 
   /**
@@ -306,7 +214,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
   public static function randomFieldAllowedBundles($entity_type, $bundle, $field_name, $count = NULL) {
     $allowed_values = [];
 
-    $field_info = FieldConfig::loadByName($entity_type, $bundle, $field_name);
+    $field_info = self::$entityTypeManager->getStorage('field_config')->load($entity_type . '.' . $bundle . '.' . $field_name);
     if ($field_info) {
       if ($field_info->getType() == 'entity_reference_revisions' || $field_info->getType() == 'entity_reference') {
         $allowed_values = $field_info->getSetting('handler_settings')['target_bundles'];
@@ -315,7 +223,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
 
     $allowed_values = array_keys($allowed_values);
 
-    return $count ? GeneratedContentRandom::arrayItems($allowed_values, $count) : $allowed_values;
+    return $count ? self::randomArrayItems($allowed_values, $count) : $allowed_values;
   }
 
   /**
@@ -371,7 +279,7 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
     // Note that we are asking for an item 1 level deeper because this is
     // how loadTree() calculates max depth.
     /** @var \Drupal\taxonomy\Entity\Term[] $tree */
-    $tree = \Drupal::entityTypeManager()->getStorage('taxonomy_term')->loadTree($vid, 0, $depth + 1, $load_entities);
+    $tree = self::$entityTypeManager->getStorage('taxonomy_term')->loadTree($vid, 0, $depth + 1, $load_entities);
 
     foreach ($tree as $k => $leaf) {
       if ($leaf->depth != $depth) {
@@ -506,140 +414,10 @@ class GeneratedContentHelper implements ContainerInjectionInterface {
   }
 
   /**
-   * Replace string tokens.
-   *
-   * @param string $string
-   *   String to process.
-   * @param array $replacements
-   *   Array of replacements with keys as tokens and values as replacements.
-   * @param callable $cb
-   *   Optional callback to process values before replacement. The callback
-   *   receives a value as it was passed in $replacements and must return
-   *   a value. If not provided, a value from $replacements will be used as-is.
-   * @param string $beginToken
-   *   Optional string to define token beginning boundary. Defaults to '{'.
-   * @param string $endToken
-   *   Optional string to define token ending boundary. Defaults to '}'.
-   *
-   * @return string
-   *   String with replaced tokens.
+   * Create an image and store it as a managed file.
    */
-  public static function replaceTokens($string, array $replacements, callable $cb = NULL, $beginToken = '{', $endToken = '}') {
-    foreach ($replacements as $k => $v) {
-      $token_name = $beginToken . $k . $endToken;
-      if ($cb && is_callable($cb)) {
-        $v = call_user_func($cb, $v);
-      }
-      if (is_scalar($v)) {
-        $string = strtr($string, [$token_name => $v]);
-      }
-    }
-
-    return $string;
-  }
-
-  /**
-   * Intersect arrays by column.
-   *
-   * @param string $column
-   *   Column name.
-   * @param ...
-   *   Variable number of arrays.
-   *
-   * @return array
-   *   Array of intersected values.
-   */
-  protected static function arrayIntersectColumn($column) {
-    $arrays = func_get_args();
-    array_shift($arrays);
-
-    if (count($arrays) < 1) {
-      throw new \Exception('At least one array argument is required');
-    }
-
-    foreach ($arrays as $k => $array) {
-      if (!is_array($array)) {
-        throw new \Exception(sprintf('Argument %s is not an array', $k + 1));
-      }
-    }
-
-    $carry = array_shift($arrays);
-    foreach ($arrays as $k => $array) {
-      $carry_column = self::arrayColumn($carry, $column);
-      $array_column = self::arrayColumn($array, $column);
-      $column_values = array_intersect($carry_column, $array_column);
-
-      $carry = array_filter($array, function ($item) use ($column, $column_values) {
-        $value = self::extractProperty($item, $column);
-
-        return $value && in_array($value, $column_values);
-      });
-    }
-
-    return $carry;
-  }
-
-  /**
-   * Portable array_column with support for methods.
-   */
-  protected static function arrayColumn(array $array, $key) {
-    if (!is_scalar($key)) {
-      throw new \Exception('Specified key is not scalar');
-    }
-
-    return array_map(function ($item) use ($key) {
-      return self::extractProperty($item, $key);
-    }, $array);
-  }
-
-  /**
-   * Helper to extract property.
-   *
-   * Note that this helper supports extracting values from simple methods.
-   *
-   * @param mixed $item
-   *   Array or object.
-   * @param string $key
-   *   Array key or object property or method.
-   *
-   * @return mixed|null
-   *   For arrays - value at specified key.
-   *   For objects - value of the specified property or returned value of the
-   *   method.
-   *
-   * @throws \Exception
-   *   If key is not scalar.
-   *   If item is not an array or an object.
-   *   If item is an object, but does not have a property or a method with
-   *   specified name.
-   *   If item is an array and does not have an element with specified key.
-   */
-  protected static function extractProperty($item, $key) {
-    if (!is_scalar($key)) {
-      throw new \Exception('Specified key is not scalar');
-    }
-
-    if (!is_object($item) && !is_array($item)) {
-      throw new \Exception(sprintf('Item with key "%s" must be an object or an array', $key));
-    }
-
-    if (is_object($item)) {
-      if (method_exists($item, $key)) {
-        return $item->{$key}();
-      }
-      elseif (property_exists($item, $key)) {
-        return $item->{$key};
-      }
-      throw new \Exception(sprintf('Key "%s" is not a property or a method of an object', $key));
-    }
-    elseif (is_array($item)) {
-      if (isset($item[$key])) {
-        return $item[$key];
-      }
-      throw new \Exception(sprintf('Key "%s" does not exist in array', $key));
-    }
-
-    return NULL;
+  public static function createImage($options = []) {
+    return self::$assetGenerator->createImage($options);
   }
 
 }
